Break out SchemaGenerator to schema.py to logically separate it.
Add doc generation endpoints to top-level urls, as it's not hardware-specific. Add versioning support to auto-docs.
